[localevaluation] try to be on the safe side when computing orders

+++ dune/gdt/localevaluation/elliptic.hh
@@ -108,8 +108,8 @@ public:
                const Stuff::LocalfunctionSetInterface< E, D, d, R, rA, rCA >& ansatzBase) const
   {
     return localFunction.order()
-        + std::max(int(testBase.order() - 1), 0)
-        + std::max(int(ansatzBase.order() - 1), 0);
+        + std::max(ssize_t(testBase.order()) - 1, ssize_t(0))
+        + std::max(ssize_t(ansatzBase.order()) - 1, ssize_t(0));
   }
 
   /**
